refactor(events): extract filter resolver base
## Context

Charge and billing segment filter resolvers share the same raw event combination matching and target bucket merge logic.

## Description

Add a base billing period filter resolver and move shared event matching behavior into it. Make the charge and billing segment resolvers inherit from the base class while keeping their target-specific behavior in each subclass.
